factor positron create env api implementation into separate functions
